+++ setup.py
@@ -1,15 +1,19 @@
 #!/usr/bin/env python3
+import timesheet
 from setuptools import setup
 
 setup(
     name="timesheet",
-    version_config=True,
+    version_config={
+        "version_file": "timesheet/VERSION",
+        "count_commits_from_version_file": True,
+    },
+    python_requires=">=3.9",
+    setup_requires=["setuptools-git-versioning"],
     install_requires=[
         "click>=7.1",
         "sqlalchemy>=1.3",
     ],
-    python_requires=">=3.9",
-    setup_requires=["setuptools-git-versioning"],
     extras_require={
         "dev": [
             "black>=20.8b1",
@@ -27,8 +31,8 @@ setup(
     packages=["timesheet"],
     entry_points={
         "console_scripts": [
-            "clock = timesheet:clock",
-            "timesheet = timesheet:main",
+            "clock = timesheet.cli:clock",
+            "timesheet = timesheet.cli:main",
         ]
     },
 )
